Added basic networking for experimental server.

+++ interpreter/string2.h
@@ -1,20 +1,19 @@
 /*
- * TRE interpreter
- * Copyright (c) 2005-2008,2011 Sven Klose <pixel@copei.de>
- *
- * String-type related section.
+ * tré - Copyright (c) 2005-2008,2011 Sven Klose <pixel@copei.de>
  */
 
 #ifndef TRE_STRING_H
 #define TRE_STRING_H
 
-#define TRESTRING_LEN(x) (*(ulong*)x)
-#define TRESTRING_DATA(x) (((char*) x) + sizeof (ulong))
+#include <sys/types.h>
+
+#define TRESTRING_LEN(x) (*(size_t*)x)
+#define TRESTRING_DATA(x) (((char*) x) + sizeof (size_t))
 
 extern treptr tre_strings;
 extern treptr trestring_get (const char *string);
 extern void trestring_free (treptr);
-extern char *trestring_get_raw (ulong len);
+extern char *trestring_get_raw (size_t len);
 
 extern struct tre_sequence_type trestring_seqtype;
